PLU-275: [BTN-MIGRATION-3] Introduce info message for Postman SMS (#607)
# Problem
We want to make it clear to users that our SMS action is only free until a certain date.

# Solution
Users may not read our guidebook, so we will put this message inside our UI (where there is a higher chance of being read)

This PR introduces a new "setup message" attribute to display such messages.

![Screenshot 2024-06-03 at 5.00.04 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/52yORk4WhOvpiff9IQU6/ee00635a-43da-4f6e-a977-da233b5f3691.png)

# Tests
- Check that info message shows up when Postman SMS is selected
- Regression test: check that info message does not show up for other apps.