PLU-275: [BTN-MIGRATION-4] Enable users to add Postman SMS connections (#625)
# Problem
We need to enable users to use their own agencies' sender ID when sending SMSes.

# Solution
Instead of using a Plumber default campaign, we will make the Postman SMS app support user added connections instead. This will enable them to use their agency's campaign and sender ID.

Note that we will need the user to setup the campaign template in a very specific way. We will inform them of this via the info message and link to our guide.

In the future, we can consider adding a "user-added-or-default" connection mode 

# Tests
- [TODO] Updated unit tests
- Check that I can send SMS.