[PARTIAL-2] set execution status on worker (#612)
## Changes
If action throws a `PartialStepError`, we set the status to `success` while still storing the errorDetails. Therefore, if the status of the execution step is `success` and errorDetails is not null, it's treated as partial success.

However, for test runs, we will set the execution step as `failure` even for partial failures for obvious reasons.