fix: sanitise input for excel formulas in variables (#731)
## Problem

Potential data exfiltration via CSV injection in M365, a user could have
a variable from a webhook or formsg:
`=HYPERLINK("https://google.com","test link")`, and this could be passed
as a variable to `createTableRow` or `updateTableRow` in M365-excel,
which could be malicious (e.g. phishing attempt)

## Solution
### New solution
- Check the input for blacklisted formula, then prepend with an
apostrophe, ignore if its a test input or variable, just block all to be
safe.

### Old solution
```
Note: I didn't return a JSON in the `preProcessVariable` because we treat variables as strings as they need to be combined with other variables or normal text inputs, so I don't want to increase the complexity which could cause backwards compatibility issues.

So instead,
For only create-table-row and update-table-row actions,
- Sanitised every variable by pre-pending a `'` to each of the formula to check for a list of possibly malicious formulas for excel e.g. hyperlink (formulas are case insensitive in Excel)
- Then, during the processing of the entire input (which could consist of both text input and variables), I unsanitised the input by checking for `'${formula}` and removing the leading apostrophe.
  - If the input was unsanitised, I will add a leading apostrophe to the entire input to transform it into a text in case it was a formula (starting with `=`)
  - If the input was not unsanitised, I will return it as per it is (no change)

Caveat:
The only issue with this approach is that if an user types in a text input with a leading apostrophe e.g. `'OFFSET(...)`, it will unsanitise it even though it was not a variable but this is okay because it is considered an invalid input (not a proper excel formula) and regardless, when this is detected, a leading apostrophe will be added to the entire input to transform it into a text as a safeguard.

- Added unit tests as well to confirm it works.
- Checked that no pipes have executions with any of the listed formulas with a leading apostrophe right now.
```

## Tests
- [x] Pipes with create table row and update table row still work with
formulas input manually by user
- [x] Pipes with create table row and update table row with formulas
passed as variables will be converted to text with a leading apostrophe
e.g. `=HYPERLINK("AAA")` becomes `'=HYPERLINK("AAA")`