chore: improve rich text editor performance (#739)
## Problem
Rich text editors gets unresponsive and laggy when user create a lot of
rows in MultiRow fields.

This is caused by the re-rendering of the entire react-hook-form
onChange.

## Solution

The re-rendering of the entire form is due to a useWatch hook at the
form level that listens to changes to dynamically render conditional
fields (hiddenIf). The hook has been removed in favour of a watch
function that only detects changes in related fields. (for example, if
field A depends on field B, it will only re-render if field B changes)

## What to test
- [ ] Check that hiddenIf fields still work
- [ ] Check that required fields work
- [ ] Check that form fields are saved properly.
- [ ] Lastly, check that creating multiple rows in MultiRow fields do
not result in performance degradation