PLU-320: fix: only continue if should not stop execution in branches (#746)
## Problem

Right now, if the condition fails for only continue if in a branch, it
stops the execution. But the intended outcome is that the execution
carries on for the other branches.

## Solution

Check for future branches and skip to next branch when only continue if
condition fails. Added unit tests to check it works too.

## Tests
Set up a pipe with 3 only continue if conditions:
Step 1: Trigger
Step 2: Only continue if
Step 3: Branch 1
Step 4: Only continue if
Step 5: Postman email
Step 6: Branch 2
Step 7: Only continue if
Step 8: Postman email

Set all only continue if condition to pass
- [x] Two emails should be sent out

Fail the only continue if condition for step 2
- [x] No emails should be sent

Fail the only continue if condition for step 4
- [x] Email for branch 1 is not sent
- [x] Email for branch 2 is sent

Fail the only continue if condition for step 7
- [x] Email for branch 1 is sent
- [x] Email for branch 2 is not sent

Fail the only continue if condition for both steps 4 and 7
- [x] Emails for branch 1 and 2 are not sent