PLU-311: [GSIB-1] proxy arcade url on gsib browser (#779)
## Problem
GSIB not able to load arcade videos on iframe. Results in too many redirects. Even custom domain on arcade results in a lot of CORS error and unable to load properly.

## Solution
Add https://safe.menlosecurity.com/ as a prefix to the arcard URL to use safe browsing proxy when user is on GSIB. We detect by inspecting `window.name` and looking for menlosecurity url