fix: stop deleting execution steps on step delete (#852)
## Problem
RDS CPU Utilisation spikes when user deletes a step due to the deletion
of corresponding execution steps.

## Solution
**Improvements**:

- Do not delete the execution steps as there is no impact to the pipe
when execution steps are not deleted
- Only impact is at the retrieval of test execution steps, where deleted
steps have to be filtered out. This tradeoff is deemed acceptable as the
query to delete execution steps is expensive for Pipes with many
executions.

## Tests
- [ ] Deleting steps does not affect Pipes, i.e. data for existing steps
remain and Pipe can execute properly
- [ ] Adding new steps to a Pipe works and Pipe executes properly