feat: add payment forms action to paysg integration (#870)
## Problem

The PaySG team wants to add a new action to the existing PaySG
integration. This PR implements this action.

For background, the PaySG team is implementing an integration with
FormSG so that payers can have a cohesive experience filling in a FormSG
form and then being redirected to a payment page within the same browser
session. This is currently not yet possible - the closest alternative is
to use the createPayment action in the existing PaySG integration, but
this creates a payment link that is sent out separately via email rather
than being a redirect flow.

Currently, users will need to use the "Custom HTTP request" action in
order to implement it, as a workaround. However, the team wants to make
this an action actually available on the UI.

## Solution

The action is implemented in this PR, under the PaySG integration
folder.

## Screenshots

<img width="1512" alt="image"
src="https://github.com/user-attachments/assets/9a732e3d-4e00-4297-9896-ba3fecfdb6ae"
/>

## Tests

For both PaySG staging and production
- [x] Create a new payment form in PaySG. Set up the plumber pipe with
relevant fields. Now, go to the PaySG payment link and ensure that the
payment form works as intended.
- [x] Ensure that the payment form still works, even if all the optional
fields are missing.