feat: add payment forms subscription action to PaySG integration (#884)
## Problem

The PaySG team wants to add a new action to the existing PaySG
integration. This PR implements this action.

For background, the PaySG team is implementing a feature where payers
can sign up for a subscription where they will be recurringly billed on
a monthly basis. This action allows PaySG admins to add a FormSG form in
front of their subscription signup page. (Basically, identical to the
previous PR #870 except that after submitting the FormSG form, payers
will get directed to a subscription signup page rather than a payment
page). As such, the action looks almost identical to the previous PR,
with the exception of the associated PaySG API endpoint and some of the
data fields.

## Solution

The action is implemented in this PR, under the PaySG integration
folder.

## Screenshots

<img width="849" alt="image"
src="https://github.com/user-attachments/assets/62ce3d27-8cfd-409f-be9f-f476c4221224"
/>

## Tests

For both PaySG staging and production
- [ ] Create a new payment form in PaySG. Set up the plumber pipe with
relevant fields. Now, go to the PaySG payment link and ensure that the
subscription page works as intended.
- [ ] Ensure that the payment form still works, even if all the optional
fields are missing.