PLU-448: [UI-REVAMP-APP-SELECT-8] Add m365 configuration modal (#930)
## Problem
Users need to select the only M365 connection on govt sharepoint when
creating a step, which is troublesome

## What changed?
- Check for the single M365 connection on the `ChooseAppAndEvent` modal
screen (log error if more than 1)
- Added a new `ConfigureExcelConnection` modal screen for M365 Excel app
only
- This screen should only load if the connection is not verified or
present yet for the user (new user or after deleting the connection on
Plumber end)
  - Else, just create or update the step and close the modal
- Moved connection-related constants to a central location for better
organization

## Extra details
Sanity checked that users on prod all only have exactly one M365
connection.
SQL query:
```
SELECT
	count(*),
	user_id
FROM
	connections
WHERE
	key = 'm365-excel'
	and deleted_at is null
GROUP BY
	user_id
HAVING count(*) > 1;
```

## Screenshots
When connection is not verified yet:


https://github.com/user-attachments/assets/e420447b-e7eb-4480-bed2-d5089853ad96

After connection is verified:


https://github.com/user-attachments/assets/ad426883-a1b2-4e72-a2be-c1a5a980560f




## How to test?
1. Manually change the verified status on DB to false for excel
connection or delete the connection in Plumber
2. Then, create a new step with M365 action, it should load the
`ConfigureExcelConnection` modal screen
3. After connecting, it should go to the successful screen. Clicking on
the `Ok got it` button will create or update the step, else closing it
will not.
4. After this, every M365 step should be created without loading the
connection modal screen
5. M365 connection screen can be popped open again by clicking on
reconnect but it does no special purpose, this is in case if the
connection breaks, the user can reconnect it here.