chore: use flag to show/hide reminder templates (#1105)
### TL;DR
Use the `app_toolbox_action_forEach` LD flag to decide on the frontend whether to show the "Schedule reminders" template with or without the for-each step.

### Why make this change?
As for-each is a beta release, we only want whitelisted users to be able to view, use, and test the template. All remaining users should be able to use the existing "Schedule reminders" template.

_Using the frontend check and remove when going to GA as it is unlikely that we will have to flag templates often and we do not want to add additional parameters to the getTemplates query_


### How to test?
- [ ] Login as a non-whitelisted user, should see the original Schedule reminders template
  - [ ] Verify that template can be used
- [ ] Login as whitelisted user, should see the NEW Schedule reminders template with for-each and find multiple rows
  - [ ] Verify that template can be used