chore: allow lookup for blank cells in Excel actions (#1132)
## Problem
Users are unable to search for blank values in their Excel columns using
the “Find table row” or “Find table rows” actions.
This is a key difference between Tiles and M365 Excel, which often
requires a workaround: creating a column specifically for lookup
purposes and then updating that column’s value once the workflow is
completed.


## Solution
Change the `required` property of the lookup value field from `true` to
`false`

## Tests
- [ ] Existing Excel steps that were previously set up still work
- [ ] 'Check step' is allowed when lookup value is empty
  - [ ] Find table row
  - [ ] Find table rows
- [ ] Update existing Excel step's lookup value to empty, should still
be able to fetch row(s)
  - [ ] Find table row
  - [ ] Find table rows
- [ ] New Excel steps can set lookup value to empty, and can still fetch
row(s)
  - [ ] Find table row
  - [ ] Find table rows

## Before & After Screenshots

https://github.com/user-attachments/assets/00a42934-4a12-4978-bb40-a48fb04108a1