fix: sanitize problematic whitespaces on tiptap editor (#1160)
## Problem

This is a follow up from the parent PR to attempt to fix the rest of the
input fields that may potentially face the same issue.

## Solution

This PR introduces a new utility function that removes problematic
whitespaces such as non-breaking whitespaces and zero-width whitespaces.
The RichTextEditor component will strip such characters onUpdate for all
input fields except Postman's editor field.

### To test
- unit tests have been added
- to manually test:
   1. type some text followed by a variable into an input field
1. copy the saved paramters from the db and paste into
https://www.regextester.com/109937 to check