chore: raise postman attachment size limit (#1227)
## Problem
Users are unable to upload attachments larger than 2 MB to send via
Email by Postman.
This was imposed as the following was mentioned in Postman's API
[docs](https://postman-v1.guides.gov.sg/email-api-guide/programmatic-email-api/send-email-api/attachments):
> - Each email can have up to 10 attachments.
> - Each attachment should not exceed 2MB in size.
> - The cumulative size of all attachments should not exceed 10MB.


## Solution
Following various tests, it was concluded that:
- The cumulative size of all attachments should not exceed 10MB.
- The maximum file size of a single attachment is 10MB.

With that we relax the limits on Plumber to the following:
-  Each attachment should not exceed 10MB ins size.
- The cumulative size of all attachments should not exceed 10MB.


## Bugfix
Discovered a bug that was not computing the cumulative size correctly,
allowing for files to still be selected even after exceeding the 10MB
limit.


## Tests
New limits
- [ ] Can upload and send attachments that are larger than 2MB but less
than 10MB.
- [ ] Can upload and send multiple attachments as long as cumulative
size is less than 10MB.

Bugfix
- [ ] Cannot select new file if the new file causes the cumulative size
to exceed 10MB

Existing Pipes
- [ ] Pipes with existing attachments work as per normal