chore: sanitise invalid chars from postman sms (#1226)
## TL;DR
Postman SMS actions fail due to invalid characters or characters that
are deemed expensive to send. We sanitise the Postman SMS body and
replace these characters with appropriate replacements so that the
actions do not fail and SMSes can be sent without any intervention.

See the list
[here](https://github.com/opengovsg/sgc-msg-segment-cal/blob/develop/src/libs/SmartEncodingMap.ts).

## How to test?
Connect a FormSG to the Pipe, make a submission with some text that
contains special characters, e.g., `JACK’SON YAPP’’IN`

![Screenshot 2025-09-23 at
2.05.30 PM.png](https://app.graphite.dev/user-attachments/assets/6153c10c-f93c-4d2f-84ea-6c569f10ad40.png)

- [ ] SMS should be sent with the "right single quotation mark" replaced
with a simple `'`