[COLLAB-22]: Error emails for collaborators (#1280)
## TL;DR
This PR adds the ability to send error notification emails to collaborators based on their role (editor or viewer). 

Owners and Editors of Pipes that have been shared can configure whether to send Editor(s) and/or Viewer(s) email notifications when the Pipe encounters an error.

## Other changes
* Added a "Save" button to the notifications settings so that the flow config is only updated once the frequency and collaborators to notify have been selected

## How to test?
Share a Pipe with Editor(s) and Viewer(s), trigger an error in your Pipe
- [ ] Owner should always receive email
- [ ] Editor should be cc-ed in email if "Notify editor(s)" is checked
- [ ] Viewer should be cc-ed in email if "Notify viewer(s)" is checked

Frequency setting should still be respected
- [ ] Should receive email for every error if frequency is set to always
- [ ] Should only receive email for the first error if frequency is set to once per day